Login form gives better errors for network failures.
All login failures were treated as bad credentials before. This was
because we were using Maybe for error handling. This gave misleading
messages if login failed because the network was down.

This version uses Results with Http.Errors. We now give better feedback
for network errors and timeouts.

The logic for login sessions may need to cahnge. We will need to make
logging in different when we are offline. Offline will allow us to only
require logins for sync-ing and first login. If we go that way, login
workflow will be different.